requirements.txt
amqp==2.1.4
autopep8==1.3
awsebcli==3.10.0
billiard==3.5.0.2
blessed==1.14.1
boto==2.46.1
boto3==1.4.4
botocore==1.5.19
celery==4.0.2
cement==2.8.2
colorama==0.3.7
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-braces==1.11.0
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-extensions==1.7.7
django-filter==1.0.2
django-rest-auth==0.9.1
django-storages==1.5.2
djangorestframework==3.5.4
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.1
kombu==4.0.2
oauthlib==2.0.1
olefile==0.44
pathspec==0.5.0
pep8==1.7.0
Pillow==4.0.0
psycopg2==2.7
pycodestyle==2.3.1
pycurl==7.43.0
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.10
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
termcolor==1.1.0
vine==1.1.3
wcwidth==0.1.7
websocket-client==0.40.0
